extensions/python/requirements.txt
awslambdaric
boto3
awsiotsdk
awscrt
